Refactor test and sourcing locations service structure
