+++ api/src/modules/sourcing-locations/sourcing-locations.module.ts
@@ -3,11 +3,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { SourcingLocationRepository } from 'modules/sourcing-locations/sourcing-location.repository';
 import { SourcingLocationsController } from 'modules/sourcing-locations/sourcing-locations.controller';
 import { SourcingLocationsService } from 'modules/sourcing-locations/sourcing-locations.service';
+import { MaterialsFromSourcingLocationsService } from './materials-from-sourcing-locations.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([SourcingLocationRepository])],
   controllers: [SourcingLocationsController],
-  providers: [SourcingLocationsService],
-  exports: [SourcingLocationsService],
+  providers: [SourcingLocationsService, MaterialsFromSourcingLocationsService],
+  exports: [SourcingLocationsService, MaterialsFromSourcingLocationsService],
 })
 export class SourcingLocationsModule {}
